# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Músicas Tião Carreiro e Pardinho",
        "description": "API para gerenciamento de músicas e sugestões",
        "contact": {
            "email": "tulioalves17@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor Principal"
        }
    ],
    "paths": {
...
        "/songs": {
            "get": {
                "tags": [
                    "Músicas"
                ],
                "summary": "Listar todas as músicas",
                "operationId": "ffeeacb25d362c6dc99cdeec7b5670e6",
                "responses": {
                    "200": {
                        "description": "Lista de músicas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Nome da Música"
                                            },
                                            "artist": {
                                                "type": "string",
                                                "example": "Nome do Artista"
                                            },
                                            "album": {
                                                "type": "string",
                                                "example": "Nome do Álbum"
                                            },
                                            "link": {
                                                "type": "string",
                                                "example": "https://exemplo.com/musica"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Músicas"
                ],
                "summary": "Atualizar uma música",
                "operationId": "1859ce39a6eabc1a4bbfe8ff1820ef42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Novo Nome da Música"
                                    },
                                    "artist": {
                                        "type": "string",
                                        "example": "Novo Nome do Artista"
                                    },
                                    "album": {
                                        "type": "string",
                                        "example": "Novo Nome do Álbum"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://exemplo.com/nova-musica"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Música atualizada com sucesso"
                    },
                    "404": {
                        "description": "Música não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Músicas"
                ],
                "summary": "Criar nova música",
                "operationId": "a28be4ced5c180620b197c9144c70376",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "artist",
                                    "album",
                                    "link"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nome da Música"
                                    },
                                    "artist": {
                                        "type": "string",
                                        "example": "Nome do Artista"
                                    },
                                    "album": {
                                        "type": "string",
                                        "example": "Nome do Álbum"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://exemplo.com/musica"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Música criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Song created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Nome da Música"
                                                },
                                                "artist": {
                                                    "type": "string",
                                                    "example": "Nome do Artista"
                                                },
                                                "album": {
                                                    "type": "string",
                                                    "example": "Nome do Álbum"
                                                },
                                                "link": {
                                                    "type": "string",
                                                    "example": "https://exemplo.com/musica"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Músicas"
                ],
                "summary": "Excluir uma música",
                "operationId": "0bf79638c915c4714858316f93c93f55",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Música excluída com sucesso"
                    },
                    "404": {
                        "description": "Música não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}